Added ComponentDebugger to be able to test each robot component

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystem/CollectionClaw.java
@@ -104,7 +104,7 @@ public class CollectionClaw implements Subsystem, TelemetryObservable {
         };
     }
 
-    public Action setClawPosition(int position){
+    public Action setClawPosition(double position){
         return (TelemetryPacket packet) -> {
             claw.setPosition(position);
             return false;
