+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystem/ControlClaw.java
@@ -104,7 +104,7 @@ public class ControlClaw implements Subsystem, TelemetryObservable {
         };
     }
 
-    public Action setClawPosition(int position){
+    public Action setClawPosition(double position){
         return (TelemetryPacket packet) -> {
             claw.setPosition(position);
             return false;
